third_party/blink/web_tests/external/wpt/css/css-box/margin-trim/block-container-block-end-nested-last-child-with-border-expected.txt

This is a testharness.js-based test.
[FAIL] .collapsed 2
  assert_equals: \n<item data-offset-y="68" class="collapsed"></item>\noffsetTop expected 68 but got 93
[FAIL] .collapsed 3
  assert_equals: \n<item data-offset-y="68" class="collapsed"></item>\noffsetTop expected 68 but got 368
Harness: the test ran to completion.
